Implement RawResponse object
Most times we end up pass models|arrays|... straight to views.
This really sucked because its hard testable, and you never
know if you got a model, an array or something else. So
RawResponse to the rescue. With help of this object
we can basically always be sure that we get a laravel collection.